+++ examples/transform.rs
@@ -15,28 +15,28 @@ fn main() {
 
 fn logic(game_state: &mut GameState) {
     for actor in &mut game_state.actors.values_mut() {
-        match game_state.seconds_since_startup {
+        match game_state.time_since_startup_f64 {
             x if x % (3.0 * TAU) < TAU => {
                 // reset scale and rotation
                 actor.scale = 1.0;
                 actor.rotation = UP;
                 // play with translation
-                actor.translation.x = (game_state.seconds_since_startup.cos() * 100.0) as f32;
-                actor.translation.y = (game_state.seconds_since_startup.sin() * 100.0) as f32;
+                actor.translation.x = (game_state.time_since_startup_f64.cos() * 100.0) as f32;
+                actor.translation.y = (game_state.time_since_startup_f64.sin() * 100.0) as f32;
             }
             x if x % (3.0 * TAU) < 2.0 * TAU => {
                 // reset translation and rotation
                 actor.translation = Vec2::ZERO;
                 actor.rotation = UP;
                 // play with scale
-                actor.scale = ((game_state.seconds_since_startup * 0.5).cos().abs() * 2.0) as f32;
+                actor.scale = ((game_state.time_since_startup_f64 * 0.5).cos().abs() * 2.0) as f32;
             }
             _ => {
                 // reset translation and scale
                 actor.translation = Vec2::ZERO;
                 actor.scale = 1.0;
                 // play with rotation
-                actor.rotation = game_state.seconds_since_startup as f32;
+                actor.rotation = game_state.time_since_startup_f64 as f32;
             }
         }
     }